+++ src/main/java/com/jecsamtech/kapebackend/repository/ContactoRepository.java
@@ -2,17 +2,15 @@ package com.jecsamtech.kapebackend.repository;
 
 import com.jecsamtech.kapebackend.model.Contacto;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-
 import java.util.List;
 
 @Repository
-public interface ContactoRepository extends JpaRepository<Contacto, Integer> {
+public interface ContactoRepository extends JpaRepository<Contacto, Long> {
 
     // Buscar todos los contactos hechos por un usuario específico
-    List<Contacto> findByUsuario_IdUsuario(Integer idUsuario);
+    List<Contacto> findByUsuario_IdUsuario(Long idUsuario);
 
     // Buscar contactos por correo
     List<Contacto> findByCorreo(String correo);
